Moving rendering back into the sensors
- Instead of keeping all of the sensor rendering in the SensorWindow,
  rendering has been moved into the sensors. A new structure,
  CameraWindow, has been created to handle receiving the sensor data and
  creating the imgui window.
- Renderable trait created to encapsulate rendering logic. The Window
  will eventually use this to have a list of sensor windows to iterate
  through and update without having to manually check for the sensors as
  it currently does with the camera.
