+++ src/camera.rs
@@ -1,11 +1,22 @@
 use byteorder::{LittleEndian, ReadBytesExt};
-use crossbeam::channel::Sender;
+use crossbeam::channel::{Sender, Receiver};
+use crate::window::Renderable;
+use glium::Display;
+use glium::{
+    backend::Facade,
+    texture::{ClientFormat, RawImage2d},
+    Texture2d,
+};
 use image::jpeg::JPEGDecoder;
 use image::ImageDecoder;
 use imgui::TextureId;
+use imgui::{self, im_str, Image, Ui, Window};
+use imgui_glium_renderer::Renderer;
+use std::borrow::Cow;
 use std::io::{self, Cursor, Read};
 use std::net::SocketAddr;
 use std::net::{TcpListener, TcpStream};
+use std::rc::Rc;
 
 pub struct Camera {
     sender: Sender<CameraData>,
@@ -15,13 +26,6 @@ pub struct CameraData {
     pub image_bytes: Vec<u8>,
     pub width: u32,
     pub height: u32,
-}
-
-pub struct CameraSettings {
-    pub rotation: u8,
-    pub window_width: f32,
-    pub window_height: f32,
-    pub texture_id: Option<TextureId>,
 }
 
 impl Camera {
@@ -73,6 +77,76 @@ impl Camera {
                     "camera channel disconnected",
                 )
             })?;
+        }
+    }
+}
+
+pub struct CameraWindow {
+    pub rotation: u8,
+    pub window_width: f32,
+    pub window_height: f32,
+    pub texture_id: Option<TextureId>,
+}
+
+impl CameraWindow {
+    pub fn new() -> Self {
+        Self {
+            rotation: 0,
+            window_width: 0.0,
+            window_height: 0.0,
+            texture_id: None,
+        }
+    }
+}
+
+
+impl Renderable for CameraWindow {
+    type Item = CameraData;
+
+    /// Renders the data received from the camera sensor. This currently
+    /// assumes RGB data format.
+    fn render(
+        &mut self,
+        ui: &Ui,
+        display: &Display,
+        renderer: &mut Renderer,
+        receiver: &mut Receiver<Self::Item>,
+    ) {
+        // If we've received new camera data, update the texture. We also need
+        // to check if there is an existing texture ahead of time so we can
+        // reuse the texture instead of creating a new one each time.
+        if let Some(camera_data) = receiver.try_recv().ok() {
+            let image_frame = Some(RawImage2d {
+                data: Cow::Owned(camera_data.image_bytes.clone()),
+                width: camera_data.width as u32,
+                height: camera_data.height as u32,
+                format: ClientFormat::U8U8U8,
+            })
+            .unwrap();
+            self.window_width = camera_data.width as f32;
+            self.window_height = camera_data.height as f32;
+            let gl_texture = Texture2d::new(display.get_context(), image_frame)
+                .expect("Couldn't create new texture");
+            if let Some(tex_id) = self.texture_id {
+                renderer.textures().replace(tex_id, Rc::new(gl_texture));
+            } else {
+                self.texture_id =
+                    Some(renderer.textures().insert(Rc::new(gl_texture)));
+            }
+        }
+
+        // We call this each iteration of the CameraWindow, so we need to make
+        // sure we draw the window even if we didn't receive camera data on
+        // this iteration. However, we currently do not draw a window unless
+        // we've received our first sample from the camera.
+        if let Some(tex_id) = self.texture_id {
+            let camera_dims = [self.window_width, self.window_height];
+            Window::new(im_str!("Camera")).build(ui, || {
+                Image::new(tex_id, camera_dims)
+                    .uv0([1.0, 1.0])
+                    .uv1([0.0, 0.0])
+                    .build(&ui);
+            });
         }
     }
 }
